server2/requirement.txt
fastapi 
uvicorn 
python-multipart 
python-dotenv 
openai 
pdfplumber
azure-cosmos
azure-openai